+++ x/epoching/client/cli/tx.go
@@ -5,7 +5,10 @@ import (
 	"strconv"
 	"strings"
 
+	errorsmod "cosmossdk.io/errors"
+	"cosmossdk.io/math"
 	"github.com/spf13/cobra"
+	flag "github.com/spf13/pflag"
 
 	"github.com/babylonlabs-io/babylon/app/params"
 	"github.com/babylonlabs-io/babylon/x/epoching/types"
@@ -13,7 +16,9 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/client/tx"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/version"
+	stakingcli "github.com/cosmos/cosmos-sdk/x/staking/client/cli"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 )
 
@@ -32,6 +37,7 @@ func GetTxCmd() *cobra.Command {
 		NewRedelegateCmd(),
 		NewUnbondCmd(),
 		NewCancelUnbondingCmd(),
+		NewEditValidatorCmd(),
 	)
 
 	return cmd
@@ -225,4 +231,82 @@ $ %s tx epoching unbond %s1gghjut3ccd8ay0zduzj64hwre2fxs9ldmqhffj 100%s --from m
 	flags.AddTxFlagsToCmd(cmd)
 
 	return cmd
+}
+
+// NewEditValidatorCmd returns a CLI command handler for creating a MsgWrappedEditValidator transaction.
+func NewEditValidatorCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "edit-validator",
+		Short: "edit an existing validator account",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			moniker, _ := cmd.Flags().GetString(stakingcli.FlagEditMoniker)
+			identity, _ := cmd.Flags().GetString(stakingcli.FlagIdentity)
+			website, _ := cmd.Flags().GetString(stakingcli.FlagWebsite)
+			security, _ := cmd.Flags().GetString(stakingcli.FlagSecurityContact)
+			details, _ := cmd.Flags().GetString(stakingcli.FlagDetails)
+			description := stakingtypes.NewDescription(moniker, identity, website, security, details)
+
+			var newRate *math.LegacyDec
+
+			commissionRate, _ := cmd.Flags().GetString(stakingcli.FlagCommissionRate)
+			if commissionRate != "" {
+				rate, err := math.LegacyNewDecFromStr(commissionRate)
+				if err != nil {
+					return fmt.Errorf("invalid new commission rate: %v", err)
+				}
+
+				newRate = &rate
+			}
+
+			var newMinSelfDelegation *math.Int
+
+			minSelfDelegationString, _ := cmd.Flags().GetString(stakingcli.FlagMinSelfDelegation)
+			if minSelfDelegationString != "" {
+				msb, ok := math.NewIntFromString(minSelfDelegationString)
+				if !ok {
+					return errorsmod.Wrap(sdkerrors.ErrInvalidRequest, "minimum self delegation must be a positive integer")
+				}
+
+				newMinSelfDelegation = &msb
+			}
+
+			valAddr := sdk.ValAddress(clientCtx.GetFromAddress())
+
+			msg := stakingtypes.NewMsgEditValidator(valAddr.String(), description, newRate, newMinSelfDelegation)
+			wMsg := types.NewMsgWrappedEditValidator(msg)
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), wMsg)
+		},
+	}
+
+	cmd.Flags().AddFlagSet(flagSetDescriptionEdit())
+	cmd.Flags().AddFlagSet(flagSetCommissionUpdate())
+	cmd.Flags().AddFlagSet(stakingcli.FlagSetMinSelfDelegation())
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func flagSetDescriptionEdit() *flag.FlagSet {
+	fs := flag.NewFlagSet("", flag.ContinueOnError)
+
+	fs.String(stakingcli.FlagEditMoniker, stakingtypes.DoNotModifyDesc, "The validator's name")
+	fs.String(stakingcli.FlagIdentity, stakingtypes.DoNotModifyDesc, "The (optional) identity signature (ex. UPort or Keybase)")
+	fs.String(stakingcli.FlagWebsite, stakingtypes.DoNotModifyDesc, "The validator's (optional) website")
+	fs.String(stakingcli.FlagSecurityContact, stakingtypes.DoNotModifyDesc, "The validator's (optional) security contact email")
+	fs.String(stakingcli.FlagDetails, stakingtypes.DoNotModifyDesc, "The validator's (optional) details")
+
+	return fs
+}
+
+func flagSetCommissionUpdate() *flag.FlagSet {
+	fs := flag.NewFlagSet("", flag.ContinueOnError)
+
+	fs.String(stakingcli.FlagCommissionRate, "", "The new commission rate percentage")
+
+	return fs
 }